feat: :sparkles: add modal tabs
